ETC (챌린지 미포함이나 기존 강의에 포함되어있음)
videoDetail ( add creator ),
Edit video, delete video, edit profile route protection,
view my upload videos on profile

+++ Controllers/videoController.js
@@ -34,9 +34,11 @@ export const postUpload = async (req, res) => {
   const newVideo = await Video.create({
     fileUrl: path,
     title,
-    description
+    description,
+    creator: req.user.id
   });
-
+  req.user.videos.push(newVideo.id);
+  req.user.save();
   res.redirect(routes.videoDetail(newVideo.id));
 };
 
@@ -44,8 +46,7 @@ export const videoDetail = async (req, res) => {
   const {
     params: { id }
   } = req;
-  const video = await Video.findById(id);
-
+  const video = await Video.findById(id).populate("creator");
   res.render("videoDetail", { pageTitle: video.title, video });
 };
 
@@ -55,7 +56,12 @@ export const getEditVideo = async (req, res) => {
   } = req;
   try {
     const video = await Video.findById(id);
-    res.render("editVideo", { pageTitle: `Edit ${video.title}`, video });
+
+    if (video.creator != req.user.id) {
+      throw Error();
+    } else {
+      res.render("editVideo", { pageTitle: `Edit ${video.title}`, video });
+    }
   } catch (error) {
     console.log(error);
     res.redirect(routes.home);
@@ -80,7 +86,12 @@ export const deleteVideo = async (req, res) => {
     params: { id }
   } = req;
   try {
-    await Video.findOneAndRemove({ _id: id });
+    const video = await Video.findById(id);
+    if (video.creator != req.user.id) {
+      throw Error();
+    } else {
+      await Video.findOneAndRemove({ _id: id });
+    }
   } catch (error) {
     console.log(error);
   }
